Se vuelve a la version de nico

+++ src/dao/CuentaDao.java
@@ -9,9 +9,9 @@ public interface CuentaDao {
 
 
 	public List<Cuenta> obtenerTodos();
-	public Cuenta obtenerUno(long id);
+	public Cuenta obtenerUno(int id);
 	public boolean insertar(Cuenta cuenta);
 	public boolean editar(Cuenta cuenta);
-	public boolean borrar(long id);
+	public boolean borrar(int id);
 	
 }
