+++ src/negocio/CuentaNeg.java
@@ -8,9 +8,9 @@ import entidad.Cuenta;
 public interface CuentaNeg {
 
 	public ArrayList<Cuenta> ListarCuentas();
-	public Cuenta obtenerUno(long id);
+	public Cuenta obtenerUno(int id);
 	public boolean insertar(Cuenta cuenta);
 	public boolean editar(Cuenta cat);
-	public boolean borrar(long id);
+	public boolean borrar(int id);
 	
 }
